Confirm tests for factorial
