+++ spec/solver.rb
